requirements.txt
tensorflow
keras
flask
pillow
numpy